Updated: batch controller, failed job controller, job controller

+++ Http/Controllers/Backend/Advanced/FailedJobsController.php
@@ -121,10 +121,11 @@ class FailedJobsController extends Controller
     public function listAction(Request $request, $type): JsonResponse
     {
 
-        $permission_slug = 'can-update-failed-jobs';
+        $permission_slugs = ['can-delete-failed-jobs','can-update-failed-jobs'];
+        $permission_response = Auth::user()->hasPermissions($permission_slugs);
 
-        if(!Auth::user()->hasPermission($permission_slug)) {
-            return vh_get_permission_denied_json_response($permission_slug);
+        if(isset($permission_response['success']) && $permission_response['success'] == false) {
+            return response()->json($permission_response);
         }
         try {
             $response = FailedJob::listAction($request, $type);
@@ -145,10 +146,11 @@ class FailedJobsController extends Controller
     //----------------------------------------------------------
     public function deleteList(Request $request): JsonResponse
     {
-        $permission_slug = 'can-delete-failed-jobs';
+        $permission_slugs = ['can-delete-failed-jobs','can-update-failed-jobs'];
+        $permission_response = Auth::user()->hasPermissions($permission_slugs);
 
-        if(!Auth::user()->hasPermission($permission_slug)) {
-            return vh_get_permission_denied_json_response($permission_slug);
+        if(isset($permission_response['success']) && $permission_response['success'] == false) {
+            return response()->json($permission_response);
         }
 
         try {
@@ -170,6 +172,12 @@ class FailedJobsController extends Controller
     //----------------------------------------------------------
     public function deleteItem(Request $request, $id): JsonResponse
     {
+        $permission_slugs = ['can-delete-failed-jobs','can-update-failed-jobs'];
+        $permission_response = Auth::user()->hasPermissions($permission_slugs);
+
+        if(isset($permission_response['success']) && $permission_response['success'] == false) {
+            return response()->json($permission_response);
+        }
         try {
             $response = FailedJob::deleteItem($request, $id);
         } catch (\Exception $e) {
